add visited_sites object in pii_vault.aggregate_data;
this temp obj has same structure as the one on local storage. it also
contains a 'dirty bit' flag which is set when values get updated in
vault.
every 10 mins, all entries with dirty bit set are flushed to local
storage
and dirty bit is set to 0;

every 10 hours a check is performed if the number of entries in
vault.visited_sites
has reached 100. If so, we should remove the least frequent (or
earliest) 10 entries
from the vault. Next time when a new entry arrives, and it is not in
vault, we check
local storage and if it is there we copy it to vault, increment it in
vault, and set dirty bit to 1. if it is not in local storage either, we
only need to add it in the vault;